+++ rest/src/main/java/xyz/cryptomaven/rest/repositories/AddressesRepository.java
@@ -3,5 +3,5 @@ package xyz.cryptomaven.rest.repositories;
 import xyz.cryptomaven.rest.models.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressesRepository extends JpaRepository<Address, Integer> {
+public interface AddressesRepository extends JpaRepository<Address, Long> {
 }
