+++ rest/src/main/java/xyz/cryptomaven/rest/repositories/MetadataRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel="metadata", path = "metadata")
-public interface MetadataRepository extends JpaRepository<Metadata, Integer>, JpaSpecificationExecutor<Metadata> {
+public interface MetadataRepository extends JpaRepository<Metadata, Long>, JpaSpecificationExecutor<Metadata> {
 }
