+++ rest/src/main/java/xyz/cryptomaven/rest/repositories/NftAddressRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel="nftaddress", path = "nftaddress")
-public interface NftAddressRepository extends JpaRepository<NftAddress, Integer>, JpaSpecificationExecutor<NftAddress> {
+public interface NftAddressRepository extends JpaRepository<NftAddress, Long>, JpaSpecificationExecutor<NftAddress> {
 }
