+++ rest/src/main/java/xyz/cryptomaven/rest/repositories/NftRefRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @RepositoryRestResource(collectionResourceRel="nft_ref", path = "nft_ref")
-public interface NftRefRepository extends JpaRepository<NftRef, Integer> {
+public interface NftRefRepository extends JpaRepository<NftRef, Long> {
     NftRef findByName(String name);
