+++ rest/src/main/java/xyz/cryptomaven/rest/repositories/NftRepository.java
@@ -3,5 +3,5 @@ package xyz.cryptomaven.rest.repositories;
 import xyz.cryptomaven.rest.models.Nft;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NftRepository extends JpaRepository<Nft, Integer> {
+public interface NftRepository extends JpaRepository<Nft, Long> {
 }
