+++ rest/src/main/java/xyz/cryptomaven/rest/repositories/RawTokenRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel="raw_token", path="raw_token")
-public interface RawTokenRepository extends JpaRepository<RawToken, Integer> {
+public interface RawTokenRepository extends JpaRepository<RawToken, Long> {
 }
